refactor(various): Improved the API for the Navbar Dropdown implementation
Also, removed the useNavLinks hook and moved the NavLinkConfig type to the CustomNavLink component.
