+++ src/features/navbar/components/VerticalNav.tsx
@@ -6,13 +6,12 @@ import Nav from 'react-bootstrap/Nav';
 import { useTranslation } from 'react-i18next';
 import { languages } from '../../../i18n/config';
 import { useLogoutModal } from '../hooks/useLogoutModal';
-import { useNavLinks } from '../hooks/useNavLinks';
-import { CustomNavAction } from './CustomNavLink';
+import { CustomNavAction, CustomNavLink } from './CustomNavLink';
 import { Logo } from './Logo';
 import { NavUserDetails } from './NavUserDetails';
 import { ThemeToggle } from '../../themes/ToggleSwitch';
-import { NavDropdown } from 'features/navbar/components/NavDropdown';
 import { useLocation } from 'react-router-dom';
+import { NavCategory } from './NavCategory';
 
 type Props = {
   closeVerticalNav?: () => void;
@@ -25,7 +24,6 @@ type LanguageOption = {
 
 export const VerticalNav: FC<Props> = ({ closeVerticalNav }) => {
   const { user } = useAuth();
-  const navLinks = useNavLinks();
   const { openLogoutModal } = useLogoutModal();
   const { i18n } = useTranslation();
   const location = useLocation();
@@ -41,19 +39,8 @@ export const VerticalNav: FC<Props> = ({ closeVerticalNav }) => {
 
   const defaultLanguageOption = __languageOptions.find(language => language.value === i18n.languages[0]);
 
-  const navCategoryMap = {
-    General: navLinks.filter(link => link.label === 'Directory'),
-    Management: navLinks.filter(link => link.label === 'Users'),
-  };
-
-  const getDefaultCategoryBasedOnLocation = () => {
-    if (location.pathname === '/users') {
-      return 'Management';
-    }
-    if (location.pathname === '/agents') {
-      return 'General';
-    }
-    return null;
+  const associatedPathsIncludesCurrentLocation = (paths: string[]) => {
+    return paths.includes(location.pathname);
   };
 
   return (
@@ -65,11 +52,18 @@ export const VerticalNav: FC<Props> = ({ closeVerticalNav }) => {
       {user ? (
         <div className='nav-wrap w-100'>
           <Nav className='flex-column'>
-            <NavDropdown
-              navCategoryMap={navCategoryMap}
-              initiallyOpenedCategory={getDefaultCategoryBasedOnLocation()}
-              closeVerticalNav={closeVerticalNav}
-            />
+            <NavCategory title='General' isOpenByDefault={associatedPathsIncludesCurrentLocation(['/agents'])}>
+              <CustomNavLink
+                link={{ icon: 'stethoscope', label: 'Directory', path: '/agents' }}
+                handleSamePathNavigate={closeVerticalNav}
+              />
+            </NavCategory>
+            <NavCategory title='Administration' isOpenByDefault={associatedPathsIncludesCurrentLocation(['/users'])}>
+              <CustomNavLink
+                link={{ icon: 'user', label: 'Users', path: '/users' }}
+                handleSamePathNavigate={closeVerticalNav}
+              />
+            </NavCategory>
           </Nav>
           <Nav className='flex-column'>
             <div className='w-100 py-3 justify-content-md-start'>
