+++ src/features/navbar/index.ts
@@ -1,2 +1,1 @@
 export * from './components/VerticalNav';
-export * from './hooks/useNavLinks';
